Fix the last two known bugs re opening and saving new files
