+++ src/common/utils/known_files.js
@@ -1,16 +1,4 @@
 import path from 'path'
-import { knownFilesStore } from './store_hooks'
-
-export function editKnownFilePath(oldPath, newPath) {
-  const key = Object.keys(knownFilesStore.store).find(
-    (id) => path.normalize(knownFilesStore.store[id].path) == path.normalize(oldPath)
-  )
-  const file = knownFilesStore.get(key)
-  knownFilesStore.set(key, {
-    ...file,
-    path: newPath,
-  })
-}
 
 // TODO: days left in trial mode?
 export function displayFileName(filePath) {
